fix: bundle duplicated function
