+++ src/utility/replace.js
@@ -1,0 +1,9 @@
+function replace(obj, name, replacement, replacements, type) {
+    var orig = obj[name];
+    obj[name] = replacement(orig);
+    if (replacements) {
+      replacements[type].push([obj, name, orig]);
+    }
+  }
+
+module.exports = replace;
